refactor(photocard): moved out of photo grid into separate component
